refactor(archive-case-details): use single sticky View PDF button
