+++ app/shared/components/blocks/archive-case-details/meta/Meta.styles.ts
@@ -10,6 +10,7 @@ export const styles: Record<string, SxProps<Theme>> = {
       md: '2 / 4',
       lg: '2 / 5'
     },
+    gridRow: '3',
     display: 'flex',
     flexDirection: 'column',
     '& > * + *': {
@@ -17,10 +18,6 @@ export const styles: Record<string, SxProps<Theme>> = {
         xs: '24px',
         md: '32px'
       }
-    },
-    marginBottom: {
-      xs: '8px',
-      md: '40px'
     }
   },
 
@@ -44,25 +41,5 @@ export const styles: Record<string, SxProps<Theme>> = {
     fontWeight: 600,
     lineHeight: '110%',
     color: mainHexPallete.black
-  },
-
-  pdfButtonWrapper: {
-    marginTop: {
-      xs: '24px',
-      sm: '24px'
-    },
-    alignSelf: 'flex-start',
-    position: {
-      xs: 'static',
-      sm: 'sticky'
-    },
-    top: {
-      sm: '48px'
-    },
-    display: {
-      xs: 'none',
-      sm: 'block'
-    },
-    zIndex: 1
   }
 };
