+++ app/shared/components/blocks/archive-case-details/meta/Meta.tsx
@@ -1,24 +1,21 @@
 import { Box, Typography } from '@mui/material';
 
-import PdfButton from '../pdf-button/PdfButton';
 import { styles } from './Meta.styles';
 
 export type MetaLabels = {
   code: string;
   dates: string;
   sheets: string;
-  viewPdf: string;
 };
 
 export type MetaProps = {
   index: string;
   dateRange: string;
   sheetsCount?: number;
-  pdfUrl: string;
   labels: MetaLabels;
 };
 
-const Meta = ({ index, dateRange, sheetsCount, pdfUrl, labels }: Readonly<MetaProps>) => {
+const Meta = ({ index, dateRange, sheetsCount, labels }: Readonly<MetaProps>) => {
   return (
     <Box sx={styles.root} data-testid="ArchiveCaseDetails-meta">
       <Box sx={styles.metaBlock}>
@@ -37,10 +34,6 @@ const Meta = ({ index, dateRange, sheetsCount, pdfUrl, labels }: Readonly<MetaPr
           <Typography sx={styles.metaValue}>{sheetsCount}</Typography>
         </Box>
       )}
-
-      <Box sx={styles.pdfButtonWrapper} data-testid="ArchiveCaseDetails-pdfButtonSticky">
-        <PdfButton href={pdfUrl} label={labels.viewPdf} dataTestId="ArchiveCaseDetails-pdfButton" />
-      </Box>
     </Box>
   );
 };
